+++ frontend/src/app/main/ui/workspace/right_header.cljs
@@ -19,6 +19,7 @@
    [app.main.store :as st]
    [app.main.ui.components.dropdown :refer [dropdown]]
    [app.main.ui.context :as ctx]
+   [app.main.ui.dashboard.team]
    [app.main.ui.ds.buttons.icon-button :refer [icon-button*]]
    [app.main.ui.exports.assets :refer [export-progress-widget]]
    [app.main.ui.formats :as fmt]
@@ -31,39 +32,6 @@
 
 (def ref:persistence-status
   (l/derived :status refs/persistence))
-
-;; --- Persistence state Widget
-
-(mf/defc persistence-state-widget
-  {::mf/wrap [mf/memo]
-   ::mf/wrap-props false}
-  []
-  (let [status (mf/deref ref:persistence-status)
-        workspace-read-only? (mf/use-ctx ctx/workspace-read-only?)]
-    (when-not workspace-read-only?
-      [:div {:class (stl/css :persistence-status-widget)}
-       (case status
-         :pending
-         [:div {:class (stl/css :status-icon :pending-status)
-                :title (tr "workspace.header.unsaved")}
-          i/status-alert]
-
-         :saving
-         [:div {:class (stl/css :status-icon :pending-status)
-                :title (tr "workspace.header.unsaved")}
-          i/status-alert]
-
-         :saved
-         [:div {:class (stl/css :status-icon :saved-status)
-                :title (tr "workspace.header.saved")}
-          i/status-tick]
-
-         :error
-         [:div {:class (stl/css :status-icon :error-status)
-                :title "There was an error saving the data. Please refresh if this persists."}
-          i/status-wrong]
-
-         nil)])))
 
 ;; --- Zoom Widget
 
@@ -215,8 +183,6 @@
     [:div {:class (stl/css :workspace-header-right)}
      [:div {:class (stl/css :users-section)}
       [:& active-sessions]]
-
-     [:& persistence-state-widget]
 
      [:& export-progress-widget]