+++ frontend/src/app/plugins/shape.cljs
@@ -211,7 +211,7 @@
                   (u/display-not-valid :name value)
 
                   :else
-                  (st/emit! (dwsh/update-shapes [id] #(assoc % :name value))))))}
+                  (st/emit! (dw/end-rename-shape id value)))))}
 
            :blocked
            {:this true
@@ -515,9 +515,10 @@
             ;; not enumerable so there are no infinite loops
             :enumerable false
             :get (fn [self]
-                   (let [shape (u/proxy->shape self)
-                         parent-id (:parent-id shape)]
-                     (shape-proxy plugin-id (obj/get self "$file") (obj/get self "$page") parent-id)))}
+                   (let [shape (u/proxy->shape self)]
+                     (when-not (cfh/root? shape)
+                       (let [parent-id (:parent-id shape)]
+                         (shape-proxy plugin-id (obj/get self "$file") (obj/get self "$page") parent-id)))))}
 
            :parentX
            {:this true